Ajuste no README para instruções para rodar a aplicaçãoruções para rodar a aplicação

+++ src/main/java/br/com/sevenfood/product/sevenfoodproductapi/core/service/ProductService.java
@@ -1,14 +1,17 @@
 package br.com.sevenfood.product.sevenfoodproductapi.core.service;
 
+import br.com.sevenfood.product.sevenfoodproductapi.application.api.exception.ResourceFoundException;
 import br.com.sevenfood.product.sevenfoodproductapi.core.domain.Product;
 import br.com.sevenfood.product.sevenfoodproductapi.core.ports.in.product.*;
 import br.com.sevenfood.product.sevenfoodproductapi.core.ports.out.ProductRepositoryPort;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Slf4j
 @Service
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 public class ProductService implements CreateProductPort, UpdateProductPort, FindByIdProductPort, FindProductsPort, DeleteProductPort {
@@ -50,9 +53,15 @@ public class ProductService implements CreateProductPort, UpdateProductPort, Fin
     @Override
     public boolean remove(Long id) {
         try {
+            Product productById = findById(id);
+            if (productById == null) {
+                throw new ResourceFoundException("Product not found");
+            }
+
             productRepository.remove(id);
             return Boolean.TRUE;
-        } catch (Exception e) {
+        } catch (ResourceFoundException e) {
+            log.error("Erro ao remover produto: {}", e.getMessage());
             return Boolean.FALSE;
         }
     }
